Add AbstractAddBadgeStep
Refactor type hierarchy so that the optional color parameter only
applies to the generic addBadge step. Reverted previous changes
to the docs generator.
